Clippy changes.
