+++ examples/tests/test_env/stream_job.sh
@@ -1,4 +1,4 @@
-sleep 10 
+sleep 15 
 
 echo OUTPUT1
